Perform initial integration with the SDK editor
This consists of a few different parts:

1. A tauri plugin that is responsible for native communications.
2. A typescript interface that uses said plugin (`lib/sdk-editor`).
3. A compile-time annotation processor (`buildsrc`).
4. A UI setting to turn the integration on / off.

More specifically:

The primary API for communication with the SDK editor mod is the
SdkEditor class and its public methods. This class internally calls the
native tauri plugin, which does a few different things:

First of all, it's responsible for setting up a local UDP server on port
49494. It's also responsible for reading incoming packets and sending
outgoing packets, and (de)serializing them from / to corresponding
message structures (`SdkToQnMessage` and `QnToSdkMessage`).

These same exact messages structures are also defined in the SDK Editor
mod side, and are expected to be the same. There's a primitive protocol
version check to prevent incompatibilities.

In order to be able to read / send messages from javascript more easily,
the `buildsrc` library provides the `#[export_ts]` annotation. This
annotation can be used on Rust structs, enums, type aliases, and
constants, and will generate serde-compatible TypeScript definitions
for annotated types. Those definitions are currently being put in
`src/lib/types`.